FLaskAWS/hello_world/requirements.txt
arrow==1.2.2
attrs==21.4.0
aws-lambda-builders==1.18.0
aws-sam-cli==1.53.0
aws-sam-translator==1.46.0
awscli==1.25.35
binaryornot==0.4.4
boto3==1.24.35
botocore==1.27.35
certifi==2022.6.15
chardet==4.0.0
chevron==0.14.0
click==7.1.2
colorama==0.4.4
cookiecutter==2.1.1
dateparser==1.1.1
docker==4.2.2
docutils==0.16
Flask==1.1.4
flask-lambda-python36==0.1.0
idna==2.10
install==1.3.5
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==2.0.1
pyasn1==0.4.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.1
PyYAML==5.4.1
regex==2021.9.30
requests==2.25.1
rsa==4.7.2
s3transfer==0.6.0
serverlessrepo==0.1.10
six==1.16.0
text-unidecode==1.3
tomlkit==0.7.2
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.10
watchdog==2.1.2
websocket-client==1.3.3
Werkzeug==1.0.1